feat: board authentication

+++ application/controller/board/BoardDeleteController.php
@@ -15,6 +15,12 @@
     $boardRepository = new BoardRepository();
     $commentRepository = new CommentRepository();
     try {
+        $findUserId = $boardRepository->findUserIdById($boardId);
+        $userId = getToken($_COOKIE['JWT']['user']);
+        if ($findUserId != $userId) {
+            throw new Exception;
+        }
+
         $commentRepository->deleteByBoardId($boardId);
         $boardRepository->delete($boardId);
